+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/custom.py
@@ -291,7 +291,7 @@ def create_managed_disk(resource_group_name, disk_name, location=None,
                         source=None,  # pylint: disable=unused-argument
                         # below are generated internally from 'source'
                         source_blob_uri=None, source_disk=None, source_snapshot=None,
-                        source_storage_account_id=None, no_wait=False):
+                        source_storage_account_id=None, no_wait=False, tags=None):
     from azure.mgmt.compute.models import Disk, CreationData, DiskCreateOption
     location = location or get_resource_group_location(resource_group_name)
     if source_blob_uri:
@@ -310,7 +310,7 @@ def create_managed_disk(resource_group_name, disk_name, location=None,
         raise CLIError('usage error: --size-gb required to create an empty disk')
 
     disk = Disk(location, disk_size_gb=size_gb, creation_data=creation_data,
-                account_type=sku)
+                account_type=sku, tags=(tags or {}))
     client = _compute_client_factory()
     return client.disks.create_or_update(resource_group_name, disk_name, disk, raw=no_wait)
 
@@ -400,7 +400,8 @@ def grant_disk_access(resource_group_name, disk_name, duration_in_seconds):
 def create_snapshot(resource_group_name, snapshot_name, location=None, size_gb=None, sku='Standard_LRS',
                     source=None,  # pylint: disable=unused-argument
                     # below are generated internally from 'source'
-                    source_blob_uri=None, source_disk=None, source_snapshot=None, source_storage_account_id=None):
+                    source_blob_uri=None, source_disk=None, source_snapshot=None, source_storage_account_id=None,
+                    tags=None):
     from azure.mgmt.compute.models import Snapshot, CreationData, DiskCreateOption
 
     location = location or get_resource_group_location(resource_group_name)
@@ -420,7 +421,7 @@ def create_snapshot(resource_group_name, snapshot_name, location=None, size_gb=N
         raise CLIError('Please supply size for the snapshots')
 
     snapshot = Snapshot(location, disk_size_gb=size_gb, creation_data=creation_data,
-                        account_type=sku)
+                        account_type=sku, tags=(tags or {}))
     client = _compute_client_factory()
     return client.snapshots.create_or_update(resource_group_name, snapshot_name, snapshot)
 
@@ -472,7 +473,7 @@ def create_image(resource_group_name, name, os_type=None, location=None,  # pyli
                  source_virtual_machine=None,
                  os_blob_uri=None, data_blob_uris=None,
                  os_snapshot=None, data_snapshots=None,
-                 os_disk=None, data_disks=None):
+                 os_disk=None, data_disks=None, tags=None):
     from azure.mgmt.compute.models import (ImageOSDisk, ImageDataDisk, ImageStorageProfile, Image, SubResource,
                                            OperatingSystemStateTypes)
     # pylint: disable=line-too-long
@@ -503,7 +504,7 @@ def create_image(resource_group_name, name, os_type=None, location=None,  # pyli
         image_storage_profile = image_storage_profile = ImageStorageProfile(os_disk=os_disk, data_disks=all_data_disks)
         location = location or get_resource_group_location(resource_group_name)
         # pylint disable=no-member
-        image = Image(location, storage_profile=image_storage_profile)
+        image = Image(location, storage_profile=image_storage_profile, tags=(tags or {}))
 
     client = _compute_client_factory()
     return client.images.create_or_update(resource_group_name, name, image)
@@ -953,10 +954,22 @@ def set_diagnostics_extension(
         no_auto_upgrade=False):
     '''Enable diagnostics on a virtual machine
     '''
-    vm = get_vm(resource_group_name, vm_name)
+    client = _compute_client_factory()
+    vm = client.virtual_machines.get(resource_group_name, vm_name, 'instanceView')
     # pylint: disable=no-member
     is_linux_os = _detect_os_type_for_diagnostics_ext(vm.os_profile)
     vm_extension_name = _LINUX_DIAG_EXT if is_linux_os else _WINDOWS_DIAG_EXT
+    if is_linux_os:  # check incompatible version
+        exts = vm.instance_view.extensions or []
+        major_ver = extension_mappings[_LINUX_DIAG_EXT]['version'].split('.')[0]
+        if next((e for e in exts if e.name == vm_extension_name and
+                 not e.type_handler_version.startswith(major_ver + '.')), None):
+            logger.warning('There is an incompatible version of diagnostics extension installed. '
+                           'We will update it with a new version')
+            poller = client.virtual_machine_extensions.delete(resource_group_name, vm_name,
+                                                              vm_extension_name)
+            LongRunningOperation()(poller)
+
     return set_extension(resource_group_name, vm_name, vm_extension_name,
                          extension_mappings[vm_extension_name]['publisher'],
                          version or extension_mappings[vm_extension_name]['version'],
@@ -971,18 +984,36 @@ def set_vmss_diagnostics_extension(
     '''Enable diagnostics on a virtual machine scale set
     '''
     client = _compute_client_factory()
-    vmss = client.virtual_machine_scale_sets.get(resource_group_name,
-                                                 vmss_name)
+    vmss = client.virtual_machine_scale_sets.get(resource_group_name, vmss_name)
     # pylint: disable=no-member
     is_linux_os = _detect_os_type_for_diagnostics_ext(vmss.virtual_machine_profile.os_profile)
     vm_extension_name = _LINUX_DIAG_EXT if is_linux_os else _WINDOWS_DIAG_EXT
-    return set_vmss_extension(resource_group_name, vmss_name, vm_extension_name,
-                              extension_mappings[vm_extension_name]['publisher'],
-                              version or extension_mappings[vm_extension_name]['version'],
-                              settings,
-                              protected_settings,
-                              no_auto_upgrade)
+    if is_linux_os and vmss.virtual_machine_profile.extension_profile:  # check incompatibles
+        exts = vmss.virtual_machine_profile.extension_profile.extensions or []
+        major_ver = extension_mappings[_LINUX_DIAG_EXT]['version'].split('.')[0]
+        # For VMSS, we don't do auto-removal like VM because there is no reliable API to wait for
+        # the removal done before we can install the newer one
+        if next((e for e in exts if e.name == _LINUX_DIAG_EXT and
+                 not e.type_handler_version.startswith(major_ver + '.')), None):
+            delete_cmd = 'az vmss extension delete -g {} --vmss-name {} -n {}'.format(
+                resource_group_name, vmss_name, vm_extension_name)
+            raise CLIError("There is an incompatible version of diagnostics extension installed. "
+                           "Please remove it by running '{}', and retry. 'az vmss update-instances'"
+                           " might be needed if with manual upgrade policy".format(delete_cmd))
 
+    poller = set_vmss_extension(resource_group_name, vmss_name, vm_extension_name,
+                                extension_mappings[vm_extension_name]['publisher'],
+                                version or extension_mappings[vm_extension_name]['version'],
+                                settings,
+                                protected_settings,
+                                no_auto_upgrade)
+
+    result = LongRunningOperation()(poller)
+    UpgradeMode = get_sdk(ResourceType.MGMT_COMPUTE, "UpgradeMode", mod='models')
+    if vmss.upgrade_policy.mode == UpgradeMode.manual:
+        poller2 = update_vmss_instances(resource_group_name, vmss_name, '*')
+        LongRunningOperation()(poller2)
+    return result
 
 # Same logic also applies on vmss
 
@@ -1075,7 +1106,15 @@ def _merge_secrets(secrets):
 
 def show_default_diagnostics_configuration(is_windows_os=False):
     '''show the default config file which defines data to be collected'''
-    return get_default_diag_config(is_windows_os)
+    public_settings = get_default_diag_config(is_windows_os)
+    # pylint: disable=line-too-long
+    protected_settings_info = json.dumps({
+        'storageAccountName': "__STORAGE_ACCOUNT_NAME__",
+        # LAD and WAD are not consistent on sas token format. Call it out here
+        "storageAccountSasToken": "__SAS_TOKEN_{}__".format("WITH_LEADING_QUESTION_MARK" if is_windows_os else "WITHOUT_LEADING_QUESTION_MARK")
+    }, indent=2)
+    logger.warning('Protected settings with storage account info is required to work with the default configurations, e.g. \n' + protected_settings_info)
+    return public_settings
 
 
 def vm_show_nic(resource_group_name, vm_name, nic):