+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -14,8 +14,8 @@ unreleased
 * vmss: update arg description for 'vmss delete-instances --instance-ids'
 * vmss: hide arg 'vmss show --ids', which is not supposed to work because of 'instance-id' arg
 * BC: vmss list-instance-connection-info: include instance IDs in the output
-
-
+* vm/vmss diagnostics: provide protected settings samples, handle extension major version upgrade, etc.
+* disk/snapshot/image: expose '--tags' in the create command
 
 2.0.6 (2017-05-09)
 ++++++++++++++++++